requirements.txt
Django==5.1.2
django_allauth==65.7.0
django_environ==0.12.0
django_import_export==4.2.1
